# config/swaggerDocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Medoc Backend API",
    "version": "1.0.0",
    "description": "API документация для Medoc Backend"
  },
  "servers": [
    {
      "url": "https://medoc.vastness.ru"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/documents/contractors": {
      "get": {
        "tags": [
          "Document"
        ],
        "summary": "Получение списка контрагентов",
        "description": "Возвращает список контрагентов, которым клиника отправляла документы.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Допустимые значения: [10, 20, 30, 40, 50]"
          },
          {
            "name": "recipientName",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Поиск по ФИО"
          },
          {
            "name": "recipientPhoneNumber",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Поиск по телефону"
          },
          {
            "name": "consentToEDO",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "Фильтрация по подписанному «Согласию на ЭДО»"
          }
        ],
        "responses": {
          "200": {
            "description": "Список контрагентов",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contractors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "recipientName": {
                            "type": "string"
                          },
                          "recipientPhoneNumber": {
                            "type": "string"
                          },
                          "signedDocumentsCount": {
                            "type": "number"
                          },
                          "consentToEDO": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "page": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректное значение limit"
          },
          "403": {
            "description": "Клиника не авторизована"
          },
          "500": {
            "description": "Ошибка сервера"
          }
        }
      }
    },
...
    "/api/contractors": {
      "post": {
        "tags": [
          "Contractor"
        ],
        "summary": "Создание контрагента",
        "description": "Создаёт нового контрагента для авторизованной клиники.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "Иван"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Петров"
                  },
                  "fathersName": {
                    "type": "string",
                    "example": "Сергеевич"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "79999999999"
                  }
                },
                "required": [
                  "firstName",
                  "lastName",
                  "phoneNumber"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Контрагент успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "contractor": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "firstName": {
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        },
                        "fathersName": {
                          "type": "string"
                        },
                        "phoneNumber": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Контрагент с таким номером телефона уже существует ИЛИ не переданы необходимые поля"
          },
          "403": {
            "description": "Клиника не авторизована"
          },
          "500": {
            "description": "Ошибка при создании контрагента"
          }
        }
      }
    },
    "/api/contractors/{contractorId}": {
      "delete": {
        "tags": [
          "Contractor"
        ],
        "summary": "Удаление контрагента",
        "description": "Удаляет контрагента, если у него нет связанных документов (пустой documentIds).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "contractorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Идентификатор контрагента (Contractor._id)"
          }
        ],
        "responses": {
          "200": {
            "description": "Контрагент успешно удалён"
          },
          "400": {
            "description": "Контрагента нельзя удалить, если у него есть документы"
          },
          "403": {
            "description": "Клиника не авторизована"
          },
          "404": {
            "description": "Контрагент не найден"
          },
          "500": {
            "description": "Ошибка сервера"
          }
        }
      }
    },
    "/api/contractors/{contractorId}/documents": {
      "get": {
        "tags": [
          "Document"
        ],
        "summary": "Получение документов контрагента",
        "description": "Возвращает все документы, связанные с указанным контрагентом, принадлежащим текущей клинике.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "contractorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID контрагента (Contractor)"
          }
        ],
        "responses": {
          "200": {
            "description": "Список документов контрагента",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contractor": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "firstName": {
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        },
                        "fathersName": {
                          "type": "string"
                        },
                        "phoneNumber": {
                          "type": "string"
                        }
                      }
                    },
                    "documents": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "fileUrl": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Контрагент не найден"
          },
          "500": {
            "description": "Ошибка при получении документов контрагента"
          }
        }
      }
    }
  }
}